+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [unreleased]
+- added `mmc` and `correlationWithMetamodel` to `daily_submissions_performances`
+
 ## [2.1.2] - 2019-11-30
 - fix staking after recent changes to the GraphQL backend